Small cleanups

+++ src/main/kotlin/io/github/faecraft/heartshapedbox/effects/MorphineStatusEffect.kt
@@ -5,9 +5,9 @@ import net.minecraft.entity.effect.StatusEffect
 import net.minecraft.entity.effect.StatusEffectType
 
 public class MorphineStatusEffect : StatusEffect(StatusEffectType.BENEFICIAL, EFFECT_COLOR) {
-    override fun getTranslationKey(): String =HSBMain.MOD_ID + ".effect.morphine"
+    override fun getTranslationKey(): String = HSBMain.MOD_ID + ".effect.morphine"
 
     public companion object {
-        public const val EFFECT_COLOR: Int = 0x8b5f65.toInt()
+        public const val EFFECT_COLOR: Int = 0x8b5f65
     }
 }
